feat: move competition views under Season Home
