+++ apps/web/src/app/dashboard/games/[fixtureId]/boxscore/page.tsx
@@ -152,7 +152,7 @@ export default async function BoxScorePage({
   return (
     <div className="mx-auto max-w-4xl">
       <Link
-        href={`/dashboard/leagues/${leagueId}/schedule`}
+        href={`/dashboard/seasons/${fixture.seasonId}/schedule`}
         className="mb-4 inline-block text-sm text-primary hover:underline"
       >
         &larr; Back to Schedule
