+++ apps/web/src/app/dashboard/teams/[id]/games/[gameId]/live/page.tsx
@@ -39,7 +39,7 @@ export default async function LiveScoreboardPage({
   return (
     <div className="mx-auto max-w-2xl">
       <Link
-        href={`/dashboard/leagues/${team.leagueId}/schedule`}
+        href={`/dashboard/seasons/${fixture.seasonId}/schedule`}
         className="mb-4 inline-block text-sm text-primary hover:underline"
       >
         &larr; Back to Schedule
